eval/requirements1.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@d1aa558119859c4b205a324afabaecabd9ef375e
certifi==2025.1.31
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
datasets==3.5.0
diffusers==0.32.2
easyocr==1.7.2
fsspec==2024.12.0
hpsv2==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
hydra-core==1.3.2
idna==3.10
image-reward==1.5
imageio==2.37.0
Jinja2==3.1.6
Levenshtein==0.27.1
libclang==18.1.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==2.1.3
omegaconf==2.3.0
open_clip_torch==2.31.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
protobuf==3.20.3
psutil==7.0.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==13.3.2
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sentencepiece==0.2.0
setuptools==78.1.0
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.12.3
submitit==1.4.5
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tifffile==2025.3.30
timm==0.6.13
tokenizers==0.21.1
torch==2.6.0
torch-utils==0.1.2
torchmetrics==1.7.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
regex
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0